requirements.txt
asgiref==3.6.0
astroid==1.6.6
awscli==1.16.169
backports.zoneinfo==0.2.1
boto3==1.26.47
botocore==1.29.47
certifi==2022.12.7
chardet==3.0.4
Click==7.0
colorama==0.3.9
dj-database-url==0.5.0
Django==4.1.5
django-filter==22.1
django-storages==1.13.2
docutils==0.14
env==0.1.0
Flask==1.0.2
gunicorn==19.9.0
heroku==0.1.4
idna==2.8
ikp3db==1.4.1
isort==4.3.20
itsdangerous==1.1.0
jedi==0.11.1
Jinja2==2.10.1
jmespath==0.9.4
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
parso==0.1.1
pbr==5.2.1
Pillow==6.0.0
psycopg2-binary==2.8.3
pyasn1==0.4.5
pylint==1.8.1
pylint-django==0.8.0
pylint-flask==0.5
pylint-plugin-utils==0.5
python-dateutil==2.8.0
python-dotenv==0.21.0
pytz==2019.1
PyYAML==3.13
requests==2.22.0
rsa==3.4.2
s3transfer==0.6.0
six==1.12.0
sqlparse==0.4.3
stevedore==1.30.1
stripe==5.0.0
urllib3==1.25.11
virtualenv==16.2.0
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
Werkzeug==0.15.2
wrapt==1.11.1
